webhook: initial implementation
This commmit adds support for Webhooks (remote triggers from Taphome to force update of the whole Taphome context)
Webhooks can be used instead of continuous polling of Taphome API to reduce the amount of unnecessary HTTP traffic exchanged between this integration and Taphome.

How to use it:
1) add 'webhook_url' property key to taphome integration configuration area in configurations.yaml.
The value of this property specifies the Webhook URL, which may be called from outside to initiate Taphome integration refresh.
If the 'webhook_url' property is not specified, webhook feature will be disabled.

2) enable webhooks in Taphome configuration (Settings/Expose Devices/TapHome API/Enable Web Hook) and set
Web Hook URL to the following format:

http://<home_assistant_ip_address>:8123/api/webhook/<webhook_url>

where '<home_assistant_ip>' is the IP address of your home assistant instance and '<webhook_url>' is the URL
you specified in configurations.yaml

3) increase the value of 'update_interval' to some large number as it is no longer necessary

Signed-off-by: Stanislav Ruzani <stanislav.ruzani@gmail.com>